+++ lib/capybara/cuprite/driver.rb
@@ -15,6 +15,7 @@ module Capybara
       delegate %i[restart quit status_code timeout timeout= current_url title body
                   window_handles close_window switch_to_window within_window window_handle
                   back forward refresh wait_for_reload viewport_size device_pixel_ratio] => :browser
+      delegate %i[send_keys] => :active_element
       alias html body
       alias current_window_handle window_handle
       alias go_back back
@@ -68,6 +69,10 @@ module Capybara
 
       def frame_title
         evaluate_script("document.title")
+      end
+
+      def active_element
+        Node.new(self, browser.active_element)
       end
 
       def find_xpath(selector)